refactor: unify asset types (frontend + backend)
Frontend:
- Consolidate asset.ts and assets.ts into single file
- Add backward compatibility aliases (AssetType, OwnershipType, CreateAssetRequest)
- Add LegacyAsset and LegacyDivisionSummary for components using asset_type field
- Update imports in lib/api/assets.ts, hooks/useAssets.ts, AssetSheetClient.tsx, AssetDivisionForm.tsx
- Delete types/assets.ts

Backend:
- Consolidate PropertyType/PropertyOwner with AssetCategory/AssetOwnership
- PropertyType and PropertyOwner are now aliases to AssetCategory and AssetOwnership
- No changes to CaseProperty model behavior (uses same enum values)

All tests pass (frontend build success, backend 1202 passed)
